add functionality to postCreate form
