+++ frontend/src/components/postCreate/postCreate.scss
