+++ frontend/src/components/posts/Posts.tsx
@@ -4,11 +4,17 @@ import { useEffect, useState } from 'react';
 import axios from "axios";
 import PostType from "../../types/PostType";
 
+interface PostWithUser extends PostType {
+  user: {
+    username: string;
+    profileImage?: string;
+  }
+}
 
-const PostsAxios: React.FC = () => {
+const Posts: React.FC = () => {
 
     const baseurl = "http://localhost:3001" 
-    const [posts, setPosts] = useState<PostType[] | null>(null);
+    const [posts, setPosts] = useState<PostWithUser[] | null>();
 
     useEffect(() => {
       getPosts()
@@ -16,16 +22,16 @@ const PostsAxios: React.FC = () => {
 
     const getPosts = async () => {
       try {
-        const users = await axios.get(`${baseurl}/posts`);
-        console.log(users.data);
-        setPosts(users.data);
+          const posts = await axios.get(`${baseurl}/posts`);
+          console.log(posts.data);
+          setPosts(posts.data);
       } catch (error) {
-        console.error("error fetching posts:", error);
+          console.error("error fetching posts:", error);
       }
-    
-    }
+    };
 
     const mapPosts = () => {
+
       return(
       posts?.map(post => {
       
@@ -49,4 +55,4 @@ const PostsAxios: React.FC = () => {
     )
 }
 
-export default PostsAxios;
+export default Posts;
